Removed abits restriction file. Renamed restriction_v2 -> restriction.
