+++ libraries/chain/include/graphene/chain/int_from_operation_type.hpp
